# packages/schemas/openapi.v1.yaml
openapi: 3.1.0
info:
  title: MARID Control Plane API
  version: 0.3.0
  description: Tenant-scoped control plane. Execution records do not execute tools.
servers:
  - url: http://localhost:5056
paths:
...
  /v1/operator/sessions:
    post:
      summary: Start an audited Nocturne engineer session with ticket and reason
      security: [{ bearerAuth: [] }]
      x-required-scope: operator.session
      responses:
        '201': { description: Session started }
        '403': { description: Requires signed engineer class and tenant role }
  /v1/operator/sessions/{id}/end:
    post:
      summary: Close an active engineer session
      security: [{ bearerAuth: [] }]
      x-required-scope: operator.session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Session ended and audited }
        '404': { description: No active owned session }
  /v1/operator/sessions/{id}/actions:
    post:
      summary: Append an out-of-band engineer action to the session audit
      security: [{ bearerAuth: [] }]
      x-required-scope: operator.session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Action recorded }
        '404': { description: No active owned session }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT